+++ frontend/packages/dev-console/src/components/import/advanced/DeploymentConfigSection.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { CheckboxField, EnvironmentField } from '../../formik-fields';
+import { CheckboxField, EnvironmentField } from '@console/shared/src/components/formik-fields';
 import FormSection from '../section/FormSection';
 
 export interface DeploymentConfigSectionProps {
