+++ frontend/packages/dev-console/src/components/import/advanced/ScalingSection.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { NumberSpinnerField } from '../../formik-fields';
+import { NumberSpinnerField } from '@console/shared/src/components/formik-fields';
 import FormSection from '../section/FormSection';
 
 const ScalingSection: React.FC = () => {
