+++ frontend/packages/dev-console/src/components/import/builder/BuilderImageSelector.tsx
@@ -4,8 +4,8 @@ import { useField, useFormikContext, FormikValues } from 'formik';
 import { LoadingInline } from '@console/internal/components/utils';
 import { FormGroup, Alert } from '@patternfly/react-core';
 import { StarIcon } from '@patternfly/react-icons';
+import { getFieldId } from '@console/shared/src/components/formik-fields/field-utils';
 import { NormalizedBuilderImages } from '../../../utils/imagestream-utils';
-import { getFieldId } from '../../formik-fields/field-utils';
 import BuilderImageCard from './BuilderImageCard';
 import './BuilderImageSelector.scss';
