+++ frontend/packages/dev-console/src/components/import/route/RouteCheckbox.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { CheckboxField } from '../../formik-fields';
+import { CheckboxField } from '@console/shared/src/components/formik-fields';
 
 const RouteCheckbox: React.FC = () => {
   return (
