+++ frontend/packages/dev-console/src/components/pipelines/pipeline-resource/PipelineClusterOptions.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { TextInputTypes } from '@patternfly/react-core';
-import { InputField, DroppableFileInputField } from '../../formik-fields';
+import { InputField, DroppableFileInputField } from '@console/shared/src/components/formik-fields';
 
 const PipelineClusterOptions: React.FC = () => (
   <>
