+++ frontend/packages/dev-console/src/components/pipelines/pipeline-resource/PipelineGitOptions.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { TextInputTypes } from '@patternfly/react-core';
-import { InputField } from '../../formik-fields';
+import { InputField } from '@console/shared/src/components/formik-fields';
 
 const PipelineGitOptions: React.FC = () => (
   <>
